sched: rt: Fix broken sync wakeup logic
RT's CPU selection algorithm currently picks the task's CPU, not the
waker's CPU during sync wakeups. Fix this to pick the current CPU if its
allowed for this task. Also previously it was using the lowest priority
mask to pick only from that, this shouldn't matter since for sync wakeups
(WF_SYNC), the waker task is going to sleep soon anyway even if it is
higher priority than the task being woken up.

Tests: uibench janktests on sailfish.
Bug: 24503801

Change-Id: I81f7e05d2c051128f52a379712edea994cdd9138
Signed-off-by: Joel Fernandes <joelaf@google.com>